+++ packages/typescript/ai/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @tanstack/ai
 
+## 0.2.1
+
+### Patch Changes
+
+- Fix up model names for OpenAI and release the new response APIs ([#188](https://github.com/TanStack/ai/pull/188))
+
+- fix up readmes ([#188](https://github.com/TanStack/ai/pull/188))
+
+## 0.2.0
+
+### Minor Changes
+
+- Standard schema / standard json schema support for TanStack AI ([#165](https://github.com/TanStack/ai/pull/165))
+
+## 0.1.0
+
+### Minor Changes
+
+- Split up adapters for better tree shaking into separate functionalities ([#137](https://github.com/TanStack/ai/pull/137))
+
 ## 0.0.3
 
 ### Patch Changes